+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/ChallengeController.java
@@ -2,6 +2,7 @@ package com.itachallenge.challenge.controller;
 
 import com.itachallenge.challenge.annotations.ValidGenericPattern;
 import com.itachallenge.challenge.config.PropertiesConfig;
+import com.itachallenge.challenge.document.DetailDocument;
 import com.itachallenge.challenge.dto.*;
 import com.itachallenge.challenge.exception.BadRequestException;
 import com.itachallenge.challenge.exception.JwtException;
@@ -25,6 +26,8 @@ import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 @RestController
@@ -336,5 +339,67 @@ public class ChallengeController {
     )
     public Mono<GenericResultDto<TagDto>> getAllTags() {
         return tagService.getAllTags();
+    }
+
+    @PutMapping("/challenge/{challengeId}/update")
+    @Operation(
+            operationId = "Updates an existing challenge.",
+            summary = "Updates information of a challenge.",
+            description = "Allows to update any information contained in a challenge, providing ChallengeId and new information.",
+            responses = {
+                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = ChallengeDto.class), mediaType = "application/json")}),
+                    @ApiResponse(responseCode = "400", description = "Missing or invalid authorization header."),
+                    @ApiResponse(responseCode = "404", description = "The Challenge with given Id was not found."),
+                    @ApiResponse(responseCode = "500", description = "Internal Server Error")
+            }
+    )
+    public Mono<ResponseEntity<ChallengeDto>> updateChallenge(
+            @PathVariable String challengeId, @Valid @RequestBody ChallengeCreateDto challengeFormDto){
+
+        final DateTimeFormatter CUSTOM_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDateTime localTime = LocalDateTime.now();
+        DetailDocument detail = new DetailDocument(challengeFormDto.getDescription());
+        LanguageDto languageDto = new LanguageDto(UUID.randomUUID(), challengeFormDto.getLanguage(), null);
+        languageDto.setLanguageImage("");
+        String solutionId = "d624bae4-9a43-4515-8979-801c0d6fd88c";
+
+        ChallengeDto challengeDto = new ChallengeDto();
+        challengeDto.setChallengeId(UUID.fromString(challengeId));
+        challengeDto.setTitle(challengeFormDto.getChallengeTitle());
+        challengeDto.setLevel(challengeFormDto.getLevel() != null ?
+                String.valueOf(challengeFormDto.getLevel()) : "");
+        challengeDto.setCreationDate(localTime.format(CUSTOM_FORMATTER));
+        challengeDto.setDetail(detail);
+        challengeDto.setPopularity(10);
+        challengeDto.setPercentage(0.5F);
+        challengeDto.setLanguages(Set.of(languageDto));
+        challengeDto.setSolutions(List.of(UUID.fromString(solutionId)));
+        challengeDto.setTopic(challengeFormDto.getTopic());
+        challengeDto.setTimesFavorite(10);
+        challengeDto.setTimesBookmark(10);
+
+        return Mono.just(ResponseEntity.ok(challengeDto));
+    }
+
+    @DeleteMapping("/challenges/{challengeId}/bookmarks")
+    @Operation(
+            operationId = "Remove a challenge from the User's bookmarks.",
+            summary = "Remove a challenge from bookmarks.",
+            description = "The ID Challenge sent through the URI is removed from the user's bookmarks. User Id is determined from the headers.",
+            responses = {
+                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = FavoriteDto.class), mediaType = "application/json")}),
+                    @ApiResponse(responseCode = "400", description = "Missing or invalid authorization header."),
+                    @ApiResponse(responseCode = "404", description = "The Challenge with given Id was not found."),
+                    @ApiResponse(responseCode = "500", description = "Internal Server Error")
+            }
+    )
+    public Mono<ResponseEntity<BookmarkDto>> removeChallengeFromBookmarks(
+            @PathVariable String challengeId,
+            @RequestHeader(name = "Authorization", required = false) String authHeader) {
+        return Mono.fromCallable(() -> jwtService.getUserUuIdFromAuthenticationHeader(authHeader))
+                .onErrorMap(JwtException.class, e -> new BadRequestException(e.getMessage()))
+                .flatMap(userId -> challengeService.removeChallengeFromBookmarks(challengeId, userId))
+                .doOnError(error -> log.error("Error removing challenge with id {} from bookmarks: {}", challengeId, error.getMessage()))
+                .map(ResponseEntity::ok);
     }
 }